#248 rename : 디렉터리명 오타 수정
